Tax invoice fractional values for cgst,sgst CustomProductUpdate chnage

+++ app/src/main/java/com/adrosonic/craftexchange/database/entities/realmEntities/TaxInvDetails.kt
@@ -20,7 +20,7 @@ open class TaxInvDetails : RealmObject() {
     var deliveryCharges: Long?=0
     var cgst: Long?=0
     var sgst: Long?=0
-    var finalTotalAmt: Long?=0
+    var finalTotalAmt: Double?=0.0
     var isactive: Long?=null
     var createdon: String?=""
     var modifiedon: String?=""
